refactor: move ExpressionUnwrapper to src/utils for layer compliance
The dependency cruiser check (logic-cannot-import-output) was failing
because FunctionCallAnalyzer in logic/ imported ExpressionUnwrapper
from output/. Moving to src/utils/ allows both layers to access it.
