+++ src/transpiler/output/codegen/TypeResolver.ts
@@ -10,7 +10,7 @@ import SIGNED_TYPES from "./types/SIGNED_TYPES";
 import UNSIGNED_TYPES from "./types/UNSIGNED_TYPES";
 import TYPE_WIDTH from "./types/TYPE_WIDTH";
 import TYPE_RANGES from "./types/TYPE_RANGES";
-import ExpressionUnwrapper from "./utils/ExpressionUnwrapper";
+import ExpressionUnwrapper from "../../../utils/ExpressionUnwrapper";
 
 /**
  * Internal type info tracked through postfix suffix chains.
